refactor: extract packet data logic to hook
