+++ frontend/src/components/packet/ResponseDialog.jsx
@@ -1,35 +1,27 @@
 import React from 'react';
-import { Dialog, DialogTitle, DialogContent, DialogActions, Typography, Box, TextField, Button } from '@mui/material';
+import { Dialog, DialogTitle, DialogContent, DialogActions, Typography, Button } from '@mui/material';
+import FrameDisplay from './FrameDisplay';
 
 const ResponseDialog = ({ open, response, onClose }) => (
   <Dialog open={open} onClose={onClose} fullWidth maxWidth="md">
-    <DialogTitle>TCP 패킷 응답</DialogTitle>
+    <DialogTitle>응답 상세</DialogTitle>
     <DialogContent>
       {response && (
-        <Box sx={{ mt: 2 }}>
-          <Typography variant="subtitle1" gutterBottom>응답 메시지</Typography>
-          <Typography variant="body2" color="success.main" gutterBottom>
-            {response.message}
+        <>
+          {!response.valid && (
+            <Typography color="error" sx={{ mb: 2 }}>
+              비정상적인 응답입니다.
+            </Typography>
+          )}
+          <Typography variant="subtitle1" gutterBottom>
+            요청 패킷
           </Typography>
-
-          <Typography variant="subtitle1" sx={{ mt: 2 }} gutterBottom>텍스트 응답</Typography>
-          <TextField
-            fullWidth
-            multiline
-            rows={4}
-            value={response.response_text || ''}
-            InputProps={{ readOnly: true }}
-          />
-
-          <Typography variant="subtitle1" sx={{ mt: 2 }} gutterBottom>16진수 응답</Typography>
-          <TextField
-            fullWidth
-            multiline
-            rows={4}
-            value={response.response || ''}
-            InputProps={{ readOnly: true }}
-          />
-        </Box>
+          <FrameDisplay data={response.requestData} />
+          <Typography variant="subtitle1" gutterBottom>
+            응답 프레임
+          </Typography>
+          <FrameDisplay data={response.responseData} />
+        </>
       )}
     </DialogContent>
     <DialogActions>
